+++ packages/core/src/components/SingleSelect/SingleSelect.tsx
@@ -21,6 +21,7 @@ export const SingleSelect: FC<SelectProps> & WithStyle = React.memo(
                 disabled,
                 onFocus,
                 onBlur,
+                className,
                 isSearchable,
                 ...inputProps
             } = props,
@@ -136,6 +137,7 @@ export const SingleSelect: FC<SelectProps> & WithStyle = React.memo(
                 id={`${selectId}-wrapper`}
                 {...{ variant, disabled, minWidth, fullWidth }}
                 ref={wrapperRef}
+                className={className}
                 isSearchable={isSearchable}
                 isErrorPresent={!!props.errorText}
                 onClick={toggleOptions}