+++ packages/core/src/components/Table/Body/Cell/RowActionsCell/index.ts
@@ -1,0 +1,1 @@
+export { RowActionsCell as default } from './RowActionsCell';