+++ packages/core/src/components/Table/Body/Cell/types.ts
@@ -1,11 +1,12 @@
 import { HTMLProps } from '@medly-components/utils';
-import { TableColumnConfig } from '../../types';
+import { TableColumnConfig, TableProps } from '../../types';
 
 export type TableCellStyledProps = {
     hidden?: boolean;
     frozen?: boolean;
     align?: 'left' | 'right' | 'center';
     wrapText?: boolean;
+    tableSize?: TableProps['size'];
 };
 
 export type TableCellProps = TableCellStyledProps &
@@ -16,6 +17,7 @@ export type TableCellProps = TableCellStyledProps &
         dottedFieldName: string;
         isLoading?: boolean;
         isRowClickDisabled?: boolean;
+        tableSize?: TableProps['size'];
         addColumnMaxSize: (field: string, value: number) => void;
         as?: keyof JSX.IntrinsicElements | React.ComponentType<any>;
     };