+++ packages/core/src/components/Table/Head/HeadRow/types.ts
@@ -1,4 +1,8 @@
 import { HTMLProps, WithThemeProp } from '@medly-components/utils';
 import { GridTemplateProps } from '../../types';
 
-export type Props = GridTemplateProps & HTMLProps<HTMLTableRowElement> & WithThemeProp;
+export type Props = GridTemplateProps &
+    HTMLProps<HTMLTableRowElement> &
+    WithThemeProp & {
+        showRowWithCardStyle?: boolean;
+    };